+++ src/main/java/com/thealgorithms/datastructures/graphs/FordFulkerson.java
@@ -1,4 +1,4 @@
-package com.thealgorithms.dynamicprogramming;
+package com.thealgorithms.datastructures.graphs;
 
 import java.util.LinkedList;
 import java.util.Queue;
