+++ src/main/java/com/thealgorithms/strings/RabinKarp.java
@@ -1,4 +1,4 @@
-package com.thealgorithms.others;
+package com.thealgorithms.strings;
 
 import java.util.Scanner;
